added the compile .exe for all of the programs

+++ 1SEM/two_number_calculation.c
@@ -39,7 +39,7 @@ float calculate(int a, int b, char op)
     }
     else if (op == '/')
     {
-        result = (a / b);
+        result = ((float) a / (float) b);
     }
     else
     {
